Create new folder for report generation
- Move report generation code into seperate file
- Update error logging to remove traceback
